+++ src/main/java/de/samply/annotations/ProjectConstraints.java
@@ -1,0 +1,14 @@
+package de.samply.annotations;
+
+import de.samply.project.ProjectType;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
+public @interface ProjectConstraints {
+    ProjectType[] projectTypes() default {};
+}